+++ source/state_representation/test/tests/test_state.cpp
@@ -12,20 +12,20 @@ TEST(StateTest, Constructors) {
   EXPECT_EQ(empty1.get_name(), "");
   EXPECT_TRUE(empty1.is_empty());
 
-  State empty2(StateType::JOINTSTATE);
-  EXPECT_EQ(empty2.get_type(), StateType::JOINTSTATE);
+  State empty2(StateType::JOINT_STATE);
+  EXPECT_EQ(empty2.get_type(), StateType::JOINT_STATE);
   EXPECT_EQ(empty2.get_name(), "");
   EXPECT_TRUE(empty2.is_empty());
 
-  State empty3(StateType::CARTESIANSTATE, "test", true);
-  EXPECT_EQ(empty3.get_type(), StateType::CARTESIANSTATE);
+  State empty3(StateType::CARTESIAN_STATE, "test", true);
+  EXPECT_EQ(empty3.get_type(), StateType::CARTESIAN_STATE);
   EXPECT_EQ(empty3.get_name(), "test");
   EXPECT_TRUE(empty3.is_empty());
   empty3.set_filled();
   EXPECT_FALSE(empty3.is_empty());
 
   State state(empty3);
-  EXPECT_EQ(state.get_type(), StateType::CARTESIANSTATE);
+  EXPECT_EQ(state.get_type(), StateType::CARTESIAN_STATE);
   EXPECT_EQ(state.get_name(), "test");
   EXPECT_FALSE(state.is_empty());
   state.set_empty();
@@ -59,13 +59,13 @@ TEST(StateTest, Timestamp) {
 }
 
 TEST(StateTest, Swap) {
-  State state1(StateType::CARTESIANSTATE, "cartesian", true);
+  State state1(StateType::CARTESIAN_STATE, "cartesian", true);
   State state2(StateType::STATE, "state", false);
   swap(state1, state2);
   EXPECT_EQ(state1.get_type(), StateType::STATE);
   EXPECT_EQ(state1.get_name(), "state");
   EXPECT_FALSE(state1.is_empty());
-  EXPECT_EQ(state2.get_type(), StateType::CARTESIANSTATE);
+  EXPECT_EQ(state2.get_type(), StateType::CARTESIAN_STATE);
   EXPECT_EQ(state2.get_name(), "cartesian");
   EXPECT_TRUE(state2.is_empty());
 }